Refactor and improve code quality across multiple files
- Updated constructors to use std::move for QString parameters in PaymentRequest, ServiceMenu, and Keyboard classes.
- Changed signal-slot connections to use const references for QNetworkProxy in UnmanagedWizardPage and NetworkServiceWindow.
- Fixed various layout initialization issues in HardwareWizardPage.
- Improved memory management by initializing member variables with default values in several classes.
- Renamed methods for consistency, such as fillModem_List to fillModemList in DialupConnectionWindow.
- Corrected typos in method names and comments across multiple files.
- Enhanced readability by adjusting formatting and spacing in several source files.
- Updated test cases to improve naming conventions and ensure clarity in test descriptions.

+++ apps/WatchService/src/mac_processenumerator.cpp
@@ -28,7 +28,8 @@ void ProcessEnumerator::enumerate() {
     }
 
     auto *pids = new pid_t[pidCount];
-    pidCount = static_cast<int>(proc_listallpids(pids, pidCount * sizeof(pid_t)));
+    pidCount =
+        static_cast<int>(proc_listallpids(pids, static_cast<size_t>(pidCount) * sizeof(pid_t)));
 
     for (int i = 0; i < pidCount; ++i) {
         pid_t pid = pids[i];
